feat: tidy up code
