+++ src/app/cities/page.tsx
@@ -1,8 +1,9 @@
 import React from "react";
-import { fetchCities } from "@/lib/cities";
+
 import SearchableDropdown from "../components/SearchDropdown";
 import CityPieChart from "../components/CityPieChart";
 import CityBarChart from "../components/CityBarChart";
+import { fetchCities } from "@/lib/cities";
 
 const CitiesPage = async () => {
   const cities = await fetchCities();
