+++ src/app/components/SearchDropdown.tsx
@@ -1,33 +1,31 @@
 "use client";
 
 import React, { useEffect, useRef, useState } from "react";
-import { useSetAtom } from "jotai";
+import { useAtom, useSetAtom } from "jotai";
 
+import { searchBarAtom } from "../atoms/searchBarAtom";
 import { listingsAtom } from "../atoms/listingsAtom";
 import { cityAtom } from "../atoms/cityAtom";
-import { searchBarAtom } from "../atoms/searchBarAtom";
 
 interface SearchableDropdownProps {
   options: string[];
 }
 
 const SearchableDropdown: React.FC<SearchableDropdownProps> = ({ options }) => {
-  const [city, setCity] = useState("");
-  const [isOpen, setIsOpen] = useState(false);
-  const dropdownRef = useRef<HTMLDivElement>(null);
-  const [loading, setLoading] = useState<boolean>(false);
-  const [error, setError] = useState<string>("");
+  const [{ loading }, setSearchBarAtom] = useAtom(searchBarAtom);
   const setListingsAtom = useSetAtom(listingsAtom);
+  const dropdownRef = useRef<HTMLDivElement>(null);
+  const [isOpen, setIsOpen] = useState(false);
   const setCityAtom = useSetAtom(cityAtom);
-  const setSearchBarAtom = useSetAtom(searchBarAtom);
+  const [city, setCity] = useState("");
 
   const filteredOptions = options.filter((option) =>
     option.toLowerCase().includes(city.toLowerCase())
   );
 
   const handleOptionClick = (option: string) => {
-    setCity(option);
     setIsOpen(false);
+    setCity(option);
   };
 
   const handleClickOutside = (event: MouseEvent) => {
@@ -39,6 +37,8 @@ const SearchableDropdown: React.FC<SearchableDropdownProps> = ({ options }) => {
     }
   };
 
+  // Paco! Aca justo tuve que aplicar la cleanup function que hablamos el otro día
+  // que hace las veces del willUnmount
   useEffect(() => {
     document.addEventListener("mousedown", handleClickOutside);
     return () => {
@@ -55,16 +55,21 @@ const SearchableDropdown: React.FC<SearchableDropdownProps> = ({ options }) => {
         cache: "no-store",
       });
       data = await res.json();
-      console.log({ data });
 
       if (res.ok) {
         setListingsAtom(data.listings);
         setCityAtom(data.avgPerType);
       } else {
-        setError("An error occurred");
+        setSearchBarAtom((atom) => ({
+          ...atom,
+          error: "An error occurred",
+        }));
       }
     } catch (err) {
-      setError("Failed to fetch listings");
+      setSearchBarAtom((atom) => ({
+        ...atom,
+        error: "Failed to fetch listings",
+      }));
     } finally {
       setSearchBarAtom((atom) => ({ ...atom, loading: false }));
     }
@@ -74,22 +79,22 @@ const SearchableDropdown: React.FC<SearchableDropdownProps> = ({ options }) => {
     <div className="flex space-x-4 place-content-center mb-10">
       <div ref={dropdownRef} className="relative w-64">
         <input
-          type="text"
-          value={city}
-          onChange={(e) => setCity(e.target.value)}
-          onFocus={() => setIsOpen(true)}
           className="w-full px-4 py-2 border border-gray-300 rounded-lg shadow-sm focus:outline-none focus:ring focus:ring-blue-300"
-          placeholder="Buscar ciudad..."
+          onChange={(e) => setCity(e.target.value)}
           disabled={options.length === 0}
+          onFocus={() => setIsOpen(true)}
+          placeholder="Buscar ciudad..."
+          value={city}
+          type="text"
         />
         {isOpen && (
           <ul className="absolute w-full mt-1 bg-white border border-gray-300 rounded-lg shadow-lg max-h-60 overflow-y-auto z-50">
             {filteredOptions.length > 0 ? (
               filteredOptions.map((option, index) => (
                 <li
-                  key={index}
                   className="px-4 py-2 hover:bg-blue-100 cursor-pointer"
                   onClick={() => handleOptionClick(option)}
+                  key={index}
                 >
                   {option}
                 </li>
@@ -101,10 +106,10 @@ const SearchableDropdown: React.FC<SearchableDropdownProps> = ({ options }) => {
         )}
       </div>
       <button
-        type="button"
         className="focus:outline-none text-white bg-green-700 hover:bg-green-800 font-medium rounded-lg text-sm px-5 py-2.5 me-2 mb-2 dark:bg-green-600 dark:hover:bg-green-700 dark:focus:ring-green-800"
-        onClick={() => fetchListings()}
         disabled={city === "" || loading}
+        onClick={() => fetchListings()}
+        type="button"
       >
         Buscar!
       </button>
