+++ src/app/page.tsx
@@ -1,7 +1,4 @@
-import { fetchCities } from "../lib/cities";
-
 const PulppoStats = async () => {
-  const cities = await fetchCities();
   return <div></div>;
 };
